Finalmente pasa paco. Tenia fallos en los casos de %p debido a una mala declaracion del tipo de argumento que recibia la funcion (tenia que ser de tipo void * aka uintmat_t) y en las funciones que gestionaban la simple conversion de entero a hexadecimal (tenia que ser unsigned int)

+++ milestone 1/printf/ft_printf.h
@@ -15,14 +15,15 @@
 # include <unistd.h>
 # include <stdlib.h>
 # include <stdarg.h>
+# include <stdint.h>
 
 int		ft_putnbr(long n);
 int		ft_putpositivenbr(unsigned int n);
 int		ft_putchar(char c);
 int		ft_putstr(char *str);
-int		ft_ptrtohex(unsigned long p);
-int		ft_nbrtohex(unsigned long p);
-int		ft_uppernbrtohex(unsigned long p);
+int		ft_ptrtohex(uintptr_t p);
+int		ft_nbrtohex(unsigned int p);
+int		ft_uppernbrtohex(unsigned int p);
 int		ft_hexlen(unsigned long ptr);
 char	*ft_strtoupper(char *str);
 int		ft_printf(char const *format, ...);
